Add certificate bundle
Fetching https:// URLs won't work if we don't have CA bundles.
Let's add them. (I think @sylr is the one who reminded me of
this but I'm not sure, it was a while ago :))

Also, update canihazip because it doesn't work anymore.
(I don't know if the domain has been cybersquatted by a
French bakery, or if the person who made canihazip in the
first place quit tech to become a baker in France, but
canihazip now redirects to a fine bakery😅🥐🍰)